requirements.txt
asgiref==3.2.3
astroid==2.4.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0
django-cors-headers==3.2.1
django-filter==2.0.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
mysqlclient==1.4.6
Pillow==6.2.1
PyJWT==1.7.1
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.24.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.10
whitenoise==5.0.1
wrapt==1.12.1
